Made the layout between private and public posts consistent. Moved public and private post renders into their own components
